Add high-level file I/O interface
This patch adds a high-level API for the CC3200's file-system and
file I/O interface. The structures and methods are very similar to
Rust's official file API.

In addition to the usual open, read, write, seek functionality, a
file's info structure can be retrieved and files can be removed.