Provide high-level interface to SFLASH file system
Using the FLC wrappers, this patch adds an interface that provides
access to files on the device's SFLASH file system. Where possible
the new interface is similar to the standard Rust file API. It also
offers additional methods for committing and reverting changes made
to the file system. Committing is required to make changes take
effect.