Provide an example update application
This patch adds an updater application to the examples. Copy your
update (e.g., blinky) to /update/mcuimg.bin, then flash the update
application and start the device. The updater will copy your update
to /sys/mcuimg.bin. When you restart the device, it will execute
the update. The change is permanent.