Add full-ota and full-ota-mcuimga examples
Like simple-ota, the example full-ota downloads an update over
HTTP and applies it to the device. Full-ota expects the update
to perform a self test before committing itself. The example
full-ota-mcuimga implements the update, which will either commit
itself, or reset the device to factory defaults; depending on
the boot-info status.